fix: updated import paths after file restructuring
