+++ apps/ui/app/pages/[schema]/pages/[page]/configure.vue
@@ -5,7 +5,7 @@ import { useHead } from "#app";
 
 import Container from "../../../../../../tailwind-components/app/components/Container.vue";
 import BreadCrumbs from "../../../../../../tailwind-components/app/components/BreadCrumbs.vue";
-import ConfigurablePage from "../../../../../../tailwind-components/app/components/pages/ConfigurablePage.vue";
+import ConfigurablePage from "../../../../../../tailwind-components/app/components/cms/ConfigurablePage.vue";
 import { getPage } from "../../../../../../tailwind-components/app/utils/cms";
 
 import { useSession } from "../../../../../../tailwind-components/app/composables/useSession";
